Cross node events.
Instrumentation calls to instrument the execution of WorkDescriptors
on remote nodes showing communication lines between nodes have been
added, however there is no support from the merger yet.

+++ src/core/instrumentation.hpp
@@ -258,7 +258,7 @@ inline Instrumentation::Event::ConstKVList Instrumentation::Event::getKVs () con
 
 inline unsigned int Instrumentation::Event::getDomain ( void ) const { return _ptpDomain; }
 
-inline unsigned int Instrumentation::Event::getId( void ) const { return _ptpId; }
+inline long long Instrumentation::Event::getId( void ) const { return _ptpId; }
 
 inline void Instrumentation::Event::reverseType ( )
 {
